Restrict MagAmmo Scarcity to whatever it spawned with
This most likely will be the full amount, but in the case of, say, the 4mm Bulk Box which is intended to spawn in empty, spilling hundreds of individual rounds is quite unintentional.  This way, whatever the original magazine was intending to spawn with will become its max, or the value calculated thus far, whichever is lower.

TODO: Handle recast-based magazines having less than max while still "containing more than their max" due to said recasts.